HomeMe/requirements.txt
annotated-types==0.7.0
anyio==4.12.1
asgiref==3.11.0
cachetools==6.2.4
certifi==2025.11.12
celery==5.6.2
charset-normalizer==3.4.4
crispy-bootstrap5==2025.6
Django==5.2.9
django-crispy-forms==2.5
django-dotenv==1.4.2
django-environ==0.12.0
dotenv==0.9.9
google-ai-generativelanguage==0.6.15
google-api-core==2.28.1
google-api-python-client==2.187.0
google-auth==2.45.0
google-auth-httplib2==0.3.0
google-generativeai==0.8.6
googleapis-common-protos==1.72.0
grpcio==1.76.0
grpcio-status==1.71.2
h11==0.16.0
httpcore==1.0.9
httplib2==0.31.0
httpx==0.28.1
idna==3.11
numpy==2.4.1
pgvector==0.4.2
pillow==12.1.0
proto-plus==1.27.0
protobuf==5.29.5
psycopg2-binary==2.9.11
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.12.5
pydantic_core==2.41.5
pydub==0.25.1
pyparsing==3.3.1
python-dotenv==1.2.1
python-telegram-bot==22.5
redis==7.1.1
requests==2.32.5
rsa==4.9.1
sqlparse==0.5.5
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
uritemplate==4.2.0
urllib3==2.6.2
